move all screen_time code into one src and one test file
